Python Relational Problem
This is an algorithm that compares money between three friends. It takes in their names, the amount of money each has and compares it.

It also tells you if two or three people have the same value.

For the next project, think of increasing the number of people to 10 and write a code for it.